Рефакторит код
