+++ project/src/types/index.ts
@@ -1,15 +1,4 @@
-import {OfferType} from '../const';
-
-type Offer = {
-  id: number;
-  isFavorite: boolean;
-  isPremium: boolean;
-  previewImage: string;
-  price: number;
-  rating: number;
-  title: string;
-  type: OfferType;
-}
+import Offer from './offer';
 
 export type {
   Offer
