+++ project/src/types/offer.ts
@@ -1,0 +1,14 @@
+import {OfferType} from '../const';
+
+type Offer = {
+  id: number;
+  isFavorite: boolean;
+  isPremium: boolean;
+  previewImage: string;
+  price: number;
+  rating: number;
+  title: string;
+  type: OfferType;
+}
+
+export default Offer;
